Move Boolean Operations filter to FEVV/Filters/CGAL
